requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
tqdm>=4.66.0
requests>=2.31.0
openai>=1.50.0
gradio>=4.40.0
segment-anything>=1.0
opencv-python>=4.9.0
python-dotenv>=1.0.0
